chore: refactor and move more functions from itinerary page to its utils
